Create copy of config api in a new package, and redirect old api to the new one
