+++ src/client/java/net/pneumono/pneumonocore/config_api/entries/IntegerConfigurationEntry.java
@@ -1,4 +1,4 @@
-package net.pneumono.pneumonocore.config.entries;
+package net.pneumono.pneumonocore.config_api.entries;
 
 import com.google.common.collect.ImmutableList;
 import net.minecraft.client.gui.DrawContext;
@@ -6,32 +6,36 @@ import net.minecraft.client.gui.Element;
 import net.minecraft.client.gui.Selectable;
 import net.minecraft.client.gui.widget.SliderWidget;
 import net.minecraft.text.Text;
-import net.pneumono.pneumonocore.config.AbstractConfiguration;
-import net.pneumono.pneumonocore.config.ConfigOptionsScreen;
-import net.pneumono.pneumonocore.config.ConfigsListWidget;
-import net.pneumono.pneumonocore.config.FloatConfiguration;
+import net.minecraft.util.math.MathHelper;
+import net.pneumono.pneumonocore.config_api.AbstractConfiguration;
+import net.pneumono.pneumonocore.config_api.ConfigOptionsScreen;
+import net.pneumono.pneumonocore.config_api.ConfigsListWidget;
+import net.pneumono.pneumonocore.config_api.configurations.IntegerConfiguration;
 
 import java.util.List;
 
-public class FloatConfigurationEntry extends AbstractConfigurationEntry<FloatConfiguration> {
+public class IntegerConfigurationEntry extends AbstractConfigurationEntry<IntegerConfiguration> {
     private final ConfigSliderWidget sliderWidget;
-    private float value;
+    private int value;
 
-    public FloatConfigurationEntry(AbstractConfiguration<?> abstractConfiguration, ConfigOptionsScreen parent, ConfigsListWidget widget) {
-        super((FloatConfiguration) abstractConfiguration, parent, widget);
+    public IntegerConfigurationEntry(AbstractConfiguration<?> abstractConfiguration, ConfigOptionsScreen parent, ConfigsListWidget widget) {
+        super((IntegerConfiguration) abstractConfiguration, parent, widget);
         this.value = configuration.getValue();
+        int minValue = configuration.getMinValue();
+        int maxValue = configuration.getMaxValue();
+
         this.sliderWidget = new ConfigSliderWidget((slider, configValue) -> {
             this.parent.selectedConfiguration = configuration;
             ConfigOptionsScreen.save(configuration.getModID(), configuration.getName(), configValue);
             this.widget.update();
-        }, value, 0, 0, 110, 20);
+        }, value, minValue, maxValue, 0, 0, 110, 20);
         this.sliderWidget.setValue(value);
         this.update();
     }
 
     @Override
     public void update() {
-        float newValue = configuration.getValue();
+        int newValue = configuration.getValue();
         this.value = newValue;
         this.sliderWidget.setValue(newValue);
     }
@@ -63,33 +67,41 @@ public class FloatConfigurationEntry extends AbstractConfigurationEntry<FloatCon
 
     public static class ConfigSliderWidget extends SliderWidget {
         private final SliderChangeAction action;
+        private final int min;
+        private final int max;
 
-        public ConfigSliderWidget(SliderChangeAction action, float value, int x, int y, int width, int height) {
-            super(x, y, width, height, Text.literal(Float.toString(round(value))), value);
+        public ConfigSliderWidget(SliderChangeAction action, int value, int min, int max, int x, int y, int width, int height) {
+            super(x, y, width, height, Text.literal(Integer.toString((int)Math.round(toPercentage(value, min, max)))), toPercentage(value, min, max));
             this.action = action;
+            this.min = min;
+            this.max = max;
         }
 
         @Override
         protected void updateMessage() {
-            setMessage(Text.literal(Float.toString(round(value))));
+            setMessage(Text.literal(Integer.toString(fromPercentage(value, min, max))));
         }
 
-        private static float round(double value) {
-            return Math.round(value * 100) / 100F;
+        private static int fromPercentage(double value, int min, int max) {
+            return (int)Math.round((value * (max - min)) + min);
+        }
+
+        private static double toPercentage(int value, int min, int max) {
+            return (value - min) / (double)(max - min);
         }
 
         @Override
         protected void applyValue() {
-            action.onChange(this, round(value));
+            action.onChange(this, fromPercentage(value, min, max));
         }
 
-        public void setValue(float value) {
-            this.value = value;
+        public void setValue(int value) {
+            this.value = toPercentage(MathHelper.clamp(value, min, max), min, max);
             updateMessage();
         }
     }
 
     public interface SliderChangeAction {
-        void onChange(ConfigSliderWidget widget, float value);
+        void onChange(ConfigSliderWidget widget, int value);
     }
 }
