+++ src/client/java/net/pneumono/pneumonocore/config_api/entries/NoConfigsEntry.java
@@ -1,4 +1,4 @@
-package net.pneumono.pneumonocore.config.entries;
+package net.pneumono.pneumonocore.config_api.entries;
 
 import net.minecraft.client.font.TextRenderer;
 import net.minecraft.client.gui.DrawContext;
@@ -6,25 +6,22 @@ import net.minecraft.client.gui.Element;
 import net.minecraft.client.gui.Selectable;
 import net.minecraft.text.Text;
 import net.minecraft.util.Colors;
-import net.pneumono.pneumonocore.config.ConfigCategory;
-import net.pneumono.pneumonocore.config.ConfigOptionsScreen;
+import net.pneumono.pneumonocore.config_api.ConfigOptionsScreen;
 
 import java.util.List;
 import java.util.Objects;
 
-public class CategoryTitleEntry extends AbstractConfigListWidgetEntry {
+public class NoConfigsEntry extends AbstractConfigListWidgetEntry {
     protected final ConfigOptionsScreen parent;
-    protected final ConfigCategory category;
 
-    public CategoryTitleEntry(ConfigCategory category, ConfigOptionsScreen parent) {
+    public NoConfigsEntry(ConfigOptionsScreen parent) {
         this.parent = parent;
-        this.category = category;
     }
 
     @Override
     public void render(DrawContext context, int index, int y, int x, int entryWidth, int entryHeight, int mouseX, int mouseY, boolean hovered, float tickDelta) {
         TextRenderer textRenderer = Objects.requireNonNull(this.parent.getClient()).textRenderer;
-        context.drawCenteredTextWithShadow(textRenderer, Text.translatable(category.getTranslationKey()), x + OFFSET + 31, (y + entryHeight / 2) - 2, Colors.WHITE);
+        context.drawCenteredTextWithShadow(textRenderer, Text.translatable("configs_screen.pneumonocore.no_configs"), x + OFFSET + 31, (y + entryHeight / 2) - 2, Colors.LIGHT_RED);
     }
 
     @Override
