+++ src/client/java/net/pneumono/pneumonocore/config_api/entries/StringConfigurationEntry.java
@@ -1,4 +1,4 @@
-package net.pneumono.pneumonocore.config.entries;
+package net.pneumono.pneumonocore.config_api.entries;
 
 import com.google.common.collect.ImmutableList;
 import net.minecraft.client.gui.DrawContext;
@@ -6,10 +6,10 @@ import net.minecraft.client.gui.Element;
 import net.minecraft.client.gui.Selectable;
 import net.minecraft.client.gui.widget.TextFieldWidget;
 import net.minecraft.text.Text;
-import net.pneumono.pneumonocore.config.AbstractConfiguration;
-import net.pneumono.pneumonocore.config.ConfigOptionsScreen;
-import net.pneumono.pneumonocore.config.ConfigsListWidget;
-import net.pneumono.pneumonocore.config.StringConfiguration;
+import net.pneumono.pneumonocore.config_api.AbstractConfiguration;
+import net.pneumono.pneumonocore.config_api.ConfigOptionsScreen;
+import net.pneumono.pneumonocore.config_api.ConfigsListWidget;
+import net.pneumono.pneumonocore.config_api.configurations.StringConfiguration;
 
 import java.util.List;
 import java.util.Objects;
