+++ src/client/java/net/pneumono/pneumonocore/datagen/PneumonoCoreTranslationBuilder.java
@@ -14,8 +14,8 @@ import net.minecraft.registry.tag.TagKey;
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.stat.StatType;
 import net.minecraft.util.Identifier;
-import net.pneumono.pneumonocore.config.AbstractConfiguration;
-import net.pneumono.pneumonocore.config.EnumConfiguration;
+import net.pneumono.pneumonocore.config_api.AbstractConfiguration;
+import net.pneumono.pneumonocore.config_api.configurations.EnumConfiguration;
 
 import java.io.IOException;
 import java.nio.file.Path;
@@ -108,7 +108,25 @@ public class PneumonoCoreTranslationBuilder {
         this.builder.add(config.getTooltipTranslationKey(), tooltip);
     }
 
+    @Deprecated
+    public void addConfig(net.pneumono.pneumonocore.config.AbstractConfiguration<?, ?> config, String name, String tooltip) {
+        this.builder.add(config.getTranslationKey(), name);
+        this.builder.add(config.getTooltipTranslationKey(), tooltip);
+    }
+
     public <T extends Enum<T>> void addEnumConfig(EnumConfiguration<T> config, String name, String tooltip, String... values) {
+        this.addConfig(config, name, tooltip);
+
+        T[] keys = config.getDefaultValue().getDeclaringClass().getEnumConstants();
+        if (keys.length != values.length) throw new IllegalArgumentException("The number of enum values and translation strings must match!");
+
+        for (int i = 0; i < keys.length; ++i) {
+            this.builder.add(config.getTranslationKey(keys[i].toString().toLowerCase()), values[i]);
+        }
+    }
+
+    @Deprecated
+    public <T extends Enum<T>> void addEnumConfig(net.pneumono.pneumonocore.config.EnumConfiguration<T> config, String name, String tooltip, String... values) {
         this.addConfig(config, name, tooltip);
 
         T[] keys = config.getDefaultValue().getDeclaringClass().getEnumConstants();
