+++ src/main/java/net/pneumono/pneumonocore/PneumonoCore.java
@@ -6,7 +6,9 @@ import net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents;
 import net.fabricmc.fabric.api.resource.conditions.v1.ResourceConditionType;
 import net.fabricmc.fabric.api.resource.conditions.v1.ResourceConditions;
 import net.minecraft.util.Identifier;
-import net.pneumono.pneumonocore.config.*;
+import net.pneumono.pneumonocore.config_api.registry.ConfigPayload;
+import net.pneumono.pneumonocore.config_api.registry.PlayerJoinEvent;
+import net.pneumono.pneumonocore.config_api.registry.ServerConfigCommandRegistry;
 import net.pneumono.pneumonocore.datagen.ConfigResourceCondition;
 import net.pneumono.pneumonocore.test.PneumonoCoreTestConfigs;
 import org.slf4j.Logger;
