+++ src/main/java/net/pneumono/pneumonocore/config/ConfigEnv.java
@@ -1,7 +1,17 @@
 package net.pneumono.pneumonocore.config;
 
-@SuppressWarnings("unused")
+/**
+ * @deprecated Use {@link net.pneumono.pneumonocore.config_api.enums.ConfigEnv} instead.
+ */
+@Deprecated
 public enum ConfigEnv {
     SERVER,
-    CLIENT
+    CLIENT;
+
+    public net.pneumono.pneumonocore.config_api.enums.ConfigEnv toNew() {
+        return switch (this) {
+            case SERVER -> net.pneumono.pneumonocore.config_api.enums.ConfigEnv.SERVER;
+            case CLIENT -> net.pneumono.pneumonocore.config_api.enums.ConfigEnv.CLIENT;
+        };
+    }
 }
