+++ src/main/java/net/pneumono/pneumonocore/config_api/enums/ConfigEnv.java
@@ -1,0 +1,7 @@
+package net.pneumono.pneumonocore.config_api.enums;
+
+@SuppressWarnings("unused")
+public enum ConfigEnv {
+    SERVER,
+    CLIENT
+}
