+++ src/main/java/net/pneumono/pneumonocore/config_api/registry/PlayerJoinEvent.java
@@ -1,14 +1,15 @@
-package net.pneumono.pneumonocore.config;
+package net.pneumono.pneumonocore.config_api.registry;
 
 import net.fabricmc.fabric.api.networking.v1.PacketSender;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.network.ServerPlayNetworkHandler;
+import net.pneumono.pneumonocore.config_api.ConfigApi;
 
 public class PlayerJoinEvent implements ServerPlayConnectionEvents.Join {
     @Override
     public void onPlayReady(ServerPlayNetworkHandler handler, PacketSender sender, MinecraftServer server) {
         // Send server configurations to client
-        Configs.sendS2CConfigSyncPacket(handler.getPlayer());
+        ConfigApi.sendS2CConfigSyncPacket(handler.getPlayer());
     }
 }
