+++ src/main/java/net/pneumono/pneumonocore/config_api/registry/ServerConfigCommandRegistry.java
@@ -1,34 +1,38 @@
-package net.pneumono.pneumonocore.config;
+package net.pneumono.pneumonocore.config_api.registry;
 
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.suggestion.SuggestionProvider;
 import com.mojang.brigadier.suggestion.Suggestions;
 import com.mojang.brigadier.suggestion.SuggestionsBuilder;
-import net.fabricmc.fabric.api.client.command.v2.ClientCommandRegistrationCallback;
-import net.fabricmc.fabric.api.client.command.v2.FabricClientCommandSource;
+import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
+import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.text.Text;
 import net.minecraft.util.Formatting;
+import net.pneumono.pneumonocore.config_api.AbstractConfiguration;
+import net.pneumono.pneumonocore.config_api.ConfigApi;
+import net.pneumono.pneumonocore.config_api.ModConfigurations;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
 
-import static net.fabricmc.fabric.api.client.command.v2.ClientCommandManager.argument;
-import static net.fabricmc.fabric.api.client.command.v2.ClientCommandManager.literal;
+import static net.minecraft.server.command.CommandManager.argument;
+import static net.minecraft.server.command.CommandManager.literal;
 
-public class ClientConfigCommandRegistry {
-    public static void registerClientConfigCommand() {
-        ClientCommandRegistrationCallback.EVENT.register((dispatcher, registryAccess) ->
-            dispatcher.register(literal("clientconfig")
+public class ServerConfigCommandRegistry {
+    public static void registerServerConfigCommand() {
+        CommandRegistrationCallback.EVENT.register((dispatcher, registryAccess, environment) ->
+            dispatcher.register(literal("serverconfig")
+                .requires(source -> source.hasPermissionLevel(4))
                 .then(literal("get")
                     .then(argument("modid", StringArgumentType.string())
                         .suggests(new ModIDSuggestionProvider())
                         .then(argument("config", StringArgumentType.string())
                             .suggests(new ConfigSuggestionProvider())
                             .executes(context -> {
-                                context.getSource().sendFeedback(Text.literal(getConfigValueString(
+                                context.getSource().sendMessage(Text.literal(getConfigValueString(
                                         StringArgumentType.getString(context, "modid"),
                                         StringArgumentType.getString(context, "config")
                                 )).formatted(Formatting.AQUA));
@@ -37,13 +41,13 @@ public class ClientConfigCommandRegistry {
                             })
                         )
                         .executes(context -> {
-                            context.getSource().sendFeedback(Text.literal("Configs:"));
+                            context.getSource().sendMessage(Text.literal("Configs:"));
                             List<String> configs = getAllConfigValueStrings(StringArgumentType.getString(context, "modid"));
                             if (configs.isEmpty()) {
-                                context.getSource().sendFeedback(Text.literal("   None!"));
+                                context.getSource().sendMessage(Text.literal("   None!"));
                             } else {
                                 for (String config : configs) {
-                                    context.getSource().sendFeedback(Text.literal("   " + config));
+                                    context.getSource().sendMessage(Text.literal("   " + config));
                                 }
                             }
 
@@ -57,7 +61,7 @@ public class ClientConfigCommandRegistry {
 
     public static List<String> getAllConfigValueStrings(String modID) {
         List<String> returnConfigs = new ArrayList<>();
-        ModConfigurations modConfigs = Configs.CONFIGS.get(modID);
+        ModConfigurations modConfigs = ConfigApi.getModConfigs(modID);
         if (modConfigs != null) {
             for (AbstractConfiguration<?> config : modConfigs.configurations) {
                 returnConfigs.add(config.getModID() + ":" + config.getName() + " is set to " + config.getReloadableLoadedValue(false).toString());
@@ -67,7 +71,7 @@ public class ClientConfigCommandRegistry {
     }
 
     public static String getConfigValueString(String modID, String name) {
-        ModConfigurations modConfigs = Configs.CONFIGS.get(modID);
+        ModConfigurations modConfigs = ConfigApi.getModConfigs(modID);
         if (modConfigs != null) {
             for (AbstractConfiguration<?> config : modConfigs.configurations) {
                 if (Objects.equals(config.getName(), name)) {
@@ -78,12 +82,12 @@ public class ClientConfigCommandRegistry {
         return modID + ":" + name + " does not exist!";
     }
 
-    public static class ModIDSuggestionProvider implements SuggestionProvider<FabricClientCommandSource> {
+    public static class ModIDSuggestionProvider implements SuggestionProvider<ServerCommandSource> {
         @Override
-        public CompletableFuture<Suggestions> getSuggestions(CommandContext<FabricClientCommandSource> context, SuggestionsBuilder builder) {
-            for (ModConfigurations modConfigs : Configs.CONFIGS.values()) {
-                if (modConfigs.modID.toLowerCase().startsWith(builder.getRemainingLowerCase())) {
-                    builder.suggest(modConfigs.modID);
+        public CompletableFuture<Suggestions> getSuggestions(CommandContext<ServerCommandSource> context, SuggestionsBuilder builder) {
+            for (ModConfigurations modConfigs : ConfigApi.getModConfigs()) {
+                if (modConfigs.getModID().toLowerCase().startsWith(builder.getRemainingLowerCase())) {
+                    builder.suggest(modConfigs.getModID());
                 }
             }
 
@@ -91,10 +95,10 @@ public class ClientConfigCommandRegistry {
         }
     }
 
-    public static class ConfigSuggestionProvider implements SuggestionProvider<FabricClientCommandSource> {
+    public static class ConfigSuggestionProvider implements SuggestionProvider<ServerCommandSource> {
         @Override
-        public CompletableFuture<Suggestions> getSuggestions(CommandContext<FabricClientCommandSource> context, SuggestionsBuilder builder) {
-            ModConfigurations modConfigs = Configs.CONFIGS.get(StringArgumentType.getString(context, "modid"));
+        public CompletableFuture<Suggestions> getSuggestions(CommandContext<ServerCommandSource> context, SuggestionsBuilder builder) {
+            ModConfigurations modConfigs = ConfigApi.getModConfigs(StringArgumentType.getString(context, "modid"));
             if (modConfigs != null) {
                 for (AbstractConfiguration<?> config : modConfigs.configurations) {
                     if (config.getName().toLowerCase().startsWith(builder.getRemainingLowerCase())) {
